requirements.txt
Django
psycopg2-binary
djangorestframework
django-oauth-toolkit
rest_condition